:truck: move util functions into a proper folder
